requirements.txt
django==1.8.5

djangorestframework==3.2
djangorestframework-jwt==1.7.2

django-cors-headers
django-extensions
django-filter==0.10.0

markdown==2.6.2
werkzeug==0.10.1
django-debug-toolbar
mock==1.3.0
coverage
Pillow
ipython
ipdb
mysqlclient

